+++ NEWS.md
@@ -1,4 +1,5 @@
 # ptstem 0.0.3
+* Using the tokenizers package to extract words from text
 * Add an overview vignette
 
 # ptstem 0.0.2